Extract contents from contents
