+++ src/contact.js
@@ -1,0 +1,15 @@
+export const contactPageContent = {
+  heading: "Contact Image",
+  body: `
+  <form>
+    Telephone: <br>
+    <input type = "text" name = "telephone" value = "" >
+    <br>
+    Email: <br>
+    <input type = "text" name = "email" value = "" >
+    <br><br>
+    <input type = "submit" value = "Submit" >
+    </form>
+  `,
+  image: "https://bit.ly/2Gq50eV"
+};
